lots of killable monster refactoring
