+++ src/lib/minions/data/killableMonsters/bosses/index.ts
@@ -1,0 +1,5 @@
+import gwdBosses from './gwd';
+import miscBosses from './misc';
+import wildyBosses from './wildy';
+
+export default [...gwdBosses, ...miscBosses, ...wildyBosses];
