+++ src/lib/minions/functions/reducedTimeFromKC.ts
@@ -1,5 +1,5 @@
-import { KillableMonster } from '../../killableMonsters';
 import { Time } from '../../constants';
+import { KillableMonster } from '../types';
 
 const FIVE_HOURS = Time.Hour * 5;
